requirements.txt
# Web Framework
Flask>=2.0

# Deep Learning Framework (CPU version is generally fine for inference)
tensorflow>=2.8

# Numerical Operations
numpy>=1.21

# Image Processing
Pillow>=9.0

# PDF Processing
PyMuPDF>=1.23.0

# Google Gemini API
google-generativeai>=0.4.0

# For running the notebook
opendatasets
pandas
scikit-learn
matplotlib
seaborn